Deleted test classes BackgroundTaskAction.java, CallbackClass.java, TestClass.java. Moved HAnSCallback.java Added ToDOs
