+++ src/main/java/se/isselab/HAnS/featureExtension/FeatureService.java
@@ -8,7 +8,6 @@ import com.intellij.openapi.project.ProjectManager;
 import net.minidev.json.JSONArray;
 import net.minidev.json.JSONObject;
 import org.jetbrains.annotations.TestOnly;
-import se.isselab.HAnS.HAnSCallback;
 import se.isselab.HAnS.featureLocation.FeatureLocationManager;
 import se.isselab.HAnS.featureModel.FeatureModelUtil;
 import se.isselab.HAnS.featureLocation.FeatureFileMapping;
